Speed up creation of diffs with vim.
Creating HTML diffs with vim gets very slow when the files
become larger and there are very large blocks that differ.
To speed this up, the individual differing blocks are
shrunk when they are longer than a configured number of lines.
The information about the blocks is obtained by first calling
'diff', which is extremely fast even on large files.