Recognize text/plain files
In order to properly read text files, we need to get the encoding.
As they are identified by libmagic, we simply need to pass what appears
in the MIME type to the comparator function. We do so by capturing
part of the regex using named group, and passing the resulting dictionary
to the comparison function.