Improve optional usage of external commands
The `tool_required` decorator now raises an exception when
the command cannot be found. This enables more flexible handling.
Associated Debian package is now suggested in the comment.

The list of external tools is now available through the `--list-tools`
command-line option. We also use this output to generate the Recommends field.