Revert usage of text comparator for .info until better solution is found
Comparing two uncompressed .info files directly is working with the
text comparator. But it is crashing when it is compressed with gzip (#784891).

I tried adding a new comparator for .info files which uses the info
utility to dump the whole info manual, but this was also adding
the temporary filename (where dbd extracted the file to) into
the output (and therefore always causing differences).
Trying to pipe the content to info and let it read from /dev/stdin
was also not successful yet.

I'm reverting this back to the old behavior until a proper solution
has been found.