Don't follow symlinks, but print their targets
Symbolic links can potentially also point outside of the
compared directory. To prevent opening and comparing files
from the host, only the state (target) of symlinks should
be displayed.
This also speeds up the comparison if multiple links
are pointing to the same file (like with BusyBox in
OpenWrt images).