Comparators now return a single Difference
The forest approach was often clumsy and ill-specified. Now
comparators are expected to return a single Difference, or None.

To make it easy for comparators who are producing details, a new decorator
`returns_details` will create a wrapping Difference object for free. This
was previously a side-effect of using the `binary_fallback` decorator.
This new decorator will filter None from the list of differences,
removing some boilerplate from the comparators.