Analyze md5sums in .deb and skip identical files
We trust the md5sums in the control archive of a Debian package
to be accurate. As we already identified md5sums file as such,
we now parse them to compare their content. Files in data.tar
which are identified has having the same checksum will not be
compared later on. Huge savings!