Perform comparisons of file metadata in directories using original path
We must use the name property and not the path property because the former
is always accurate for files on the filesystem while the latter is only set
after calling get_content()… which would even be wrong for things like symlinks
or devices.

That part of the code doesn't make me happy. Gives a minefield feeling.
I don't have any good idea on how to do better though.

Closes: #796202