Improve overloading of what gets compared in a container
Most often, containers need to overload the set of files that gets
compared and not the way comparisons are performed. So we introduce
a new method get_members that returns a dictionary of names and
members. The names will be used to match which file should actually be
compared.

We also split the compare method using a 'comparisons' generator.
This should also allow other less conventional extensions in the future.

This reduces some code duplication (especially visible in .changes)
and make the situation for .gzip, .bz2, and .xz more straightforward.

We also take the opportunity to remove the useless 'source' argument
in Container.compare which was uncessarily complicating the code.