Use an iterator in Container.compare
Now that we produce the list of comparisons using an iterator, it feels
pretty logic to implement Container.compare as a transformation over
the value. itertools.starmap() to the rescue!

We still need to evaluate the iterator right away (using list) because
container code currently assumes that the content of the underlying
source is available. This could probably be avoided using proper
refactoring but that will be for later.

In any cases this change should make implementation of parallel processing
pretty straightforward.