Open file descriptors feeding diff in binary mode
We are handling encoding and decoding bytes manually when we run diff. We
encode the input to UTF-8 ourselves. We thus need to tell Python that we are
using the input file descriptor in binary mode.

This is required in Python 3 because we are writing bytes and not unicode
strings.