Call tar to retrieve Tar archive metadata
We used to rely on the Python tarfile module to get a listing with the content
of Tar archives. The interface of the list() method is a bit awkward to use as
the results are always written on stdout. Replacing sys.stdout on the fly is
not thread-safe and has problems with how Python 3 handles strings.

So let's call the system `tar`. It makes the implementation more consistent with
other comparators and we get even more accurate output.

As tar is an essential package on Debian, we filter it out from the list of
Recommends.