Propagate failure of commands feeding diff
We now check the return code once we are done with a command feeding
diff. In case of failure we raise an exception using CalledProcessError.

CalledProcessError assumes it will get the raw output. So we now capture
stderr in a BytesIO, make the attribute available, and only decode its content
when required.

If the command has received SIGTERM it's probably because we terminate commands
that linger slightly too long after we are done reading the output. So we don't
want to consider this an error.  Ideally we should wait a little longer for
them to properly shut down. That will be easily doable in Python 3.3+.