Differentiate text and bytes when computing diffs
In some situation, we want to compare streams of already encoded UTF-8 bytes,
in some other we compare streams of unicode strings. To make it clear, we split
Difference.from_file into Difference.from_raw_readers and
Difference.from_text_readers.

We want to avoid an overhead when we use the unfiltered output of a command to
feed diff directly. So we assume commands output UTF-8 encoded bytes as its
true in most cases. This leaves up to filters to similarly return UTF-8 bytes.
They might have to decode their input from UTF-8 or other encoding if required.

We also rename Difference.from_unicode into Difference.from_text to keep
names in line.

This changes are required for Python 3 which clearly separates unicode strings
and bytes, but this should clearly help avoid encoding issues in the future.