Make file detection compatible with PyPI python-magic
The Python module available in Debian as python-magic is coming from
the source of file. In Python land, its name is “Magic-file-extensions”.

Meanwhile, on PyPI, there is a module named python-magic which uses
ctypes to call the system libmagic. This module has a different
interface than Magic-file-extensions.

We add the necessary compatibility code so diffoscope can eventually
be published on PyPI.