Add git-buildpackage configuration in hope it helps to get proper source tarball
git-buildpackage has 'debuild -i -I' as its default builder. This is bad for diffoscope
as we don't want the default file ignore rules as they remove .o and .a files.
We need both kind for the test suite. If they don't get included, the package will
fail to build.

So let's specify just 'debuild' as the default builder in the hope that it will prevent
future mistake.

When #765494 is implemented in dpkg, we will be able to add a sanity check in
the form of a hook for dpkg-buildpackage. But this does not exist yet.