Force ppudump to output time in UTC
ppudump will return times using the local timezone which is not ideal to
investigate files.

Sadly, it doesn't support the TZ environment variable… but it looks for
timezone definitions in the directory specified by TZDIR.

So let's set it to a non-existent directory so we get UTC output even when the
system timezone is set otherwise. The expected diff in the test suite is
updated to reflect that.