Make container class available on demand instead of bracketing
To access the content of a container, we used to require call to the open()
method. Such bracketed operation makes it difficult to implement parallel
processing. Instead, we now make container objects associated with a given file
type available through the as_container property.

The as_container property will lazily initialize an associated container
object. The reference to the container object will be removed when the
cleanup() method is called. Ressource can thus be deallocated on garbage
collection.

With this change, we can finally make Container.compare() return an iterator
that can be lazily consumed instead of an evaluated list: the various files that
need to be extracted to perform the comparisosn will be made available when
required.

File types which can be treated as containers must now define the
CONTAINER_CLASS constant to point at a Container subclass handling this
particular file type.