Analyze md5sums file before working on data.tar
We used to be clever and analyze md5sums content to recognize files to ignore
when they were compared themselves. Sadly, this cannot be made to work easily
in the context of parallel processing.

Instead, we now explicitly extract md5sums files to analyze their content
when comparing data.tar. Given control.tar.gz is pretty small, this
should not be a significant performance hit.

The new method Container.lookup_file() allows to retrieve an abitrary file
deep down several containers.

Huge thanks to Joachim Breitner for the initial implementation.