Improve how jQuery is specified
When --html-dir is used without specifying --jquery, diffoscope will look for
its presence in a list of known locations. If found, a symlink will be created
in the report directory. This makes it easy to use --html-dir in a local manner
when libjs-jquery is installed in Debian. Supporting other operating systems
should be as easy as modifying JQUERY_SYSTEM_LOCATIONS.

If jQuery cannot be found in any known system locations, JavaScript will be
disabled and a warning printed. This will still create usable reports as links
point to proper HTML pages.

Specifying --jquery=disable will always disable JavaScript.