Remove useless <span>s in HTML output
We used to output a lot of '<span class="diffchanged2">' with just one letter
followed by the exact same span for the next character if it wass still
different. Not very efficient. So let's group changed characters together.

Thanks Mike Hommey for noticing the problem. I guess more optimizations could
be done, but that'll be for another time.

Closes: #808121