Ensure that we always get path names from libarchive as str
libarchive-c will return the pathname either as str or bytes if the string
can't be decoded in the current locale. Sadly, it makes the result very hard
to predict, especially as we might encounter weird archives.

As a solution, we monkeypatch libarchive binding so we always get a str object,
escaping undecodable bytes.

Closes: #808541