Use libarchive to list files in tar and cpio archives
The output of cpio and tar is too hard to control properly, so let's
create a similar enough output using libarchive functions.

We need to monkeypatch the libarchive binding in order to add the
missing accessors.

Sadly, there's no way in libarchive API to distinguish an archive
which would have usernames and uids from one that would have only usernames.
But good enough, I guess.

Closes: #808809