Ensure files extracted from squashfs are timely deleted
We used to parse the squashfs listing to create a dictionnary holding
all File objects. It works, but it means that we keep a reference
alive for each files as long as we are processing the squashfs.
This is bad because we now remove the temporary directory when the File
is garbage collected. We thus could thus be running out of disk space
pretty quickly for large squashfs before we finished processing them entirely.

So instead, we store in the dictionnary a reference to the relevant squashfs
member class and the arguments required to instanciate it. The instanciation
will be done when a member is requested from the SquashfsContainer.