Filter symbols for IP-relative ops from ELF disassembled output
`objdump --disassemble` displays symbols+offsets for addresses
it derives from IP-relative addressing. If the offsets are different,
that results in showing a difference even if the actual code is the same.
As we care about actual differences, let's filter them from the output.

Thanks Mike Hommey for the suggestion and the test case.

Closes: #808207