Don't show a difference for identical but long diff inputs
If the input for diff hit the number of maximum lines, we would always output
a difference, even if both input were actually the same. This usually show
when comparing ELF files and creates confusing reports.

So lets compute a hash while we feed lines to diff and write it in the final
line in case we hit the limit. In case there were differences in the hidden
part, the hash will be different and will appear as such in the report.

This should make diffoscope run longer for large diff inputs, but more correct
reports worths it.