Show line ordering variation in deb md5sums
When a md5sum file differs, diffoscope was only showing the comment
that files inside the package differ.
But it can also be that all remaining files are reproducible, and
only the line ordering inside the md5sums file varies.
If the line order varies, it will now also be shown.