Reset mtimes of test directories and ignore ctimes
The mtimes of the directories were not normalized to the same
value, which had the effect that diffoscope would also output
their time difference if it was measurable. This leads to a
test failure, as this output is not expected.

On fast systems the successive mkdir calls only sometimes resulted
in different mtimes, while on slow systems this was more often the
case (like on the armhf nodes). By adding a time.sleep(1) between
the mkdir calls, I was able to reproduce the test failure every time.

Additionally the ctimes have to be ignored, as the user has no
control over it and mkdir or os.utime calls also result in changed
ctimes (which differ if Python is not fast enough).

Closes: #815171