fix sqlite3 magic recognition when using file >= 5.27
this fixes the recognition problems of sqlite3 as newer
file versions add additional information to the magic
recognition.
The new string is like:
  SQLite 3.x database, last written using SQLite version 3008011

To be compatible with both, newer and older file versions we
check if the magic_file_type has any value and starts with the expected
recognition string.