Improve diffoscope behaviour for .changes
Being able to know details about the build process is useful
when understanding why two different builds had not come up
with the same output. It may highlight which aspects of the
build environment were different.

In the particular context of comparing .changes, we know
that if all files listed are identical except the .buildinfo
file, the differences in the .buildinfo files are irrelevant.

We thus ignore .buildinfo differences if all the files
except .buildinfo files referenced in `Files` field and all
other fields of .changes are identical.