difference: Don't truncate diff output by default, and use a separate flag when doing so
Previously, max-diff-block-lines would also (on top of altering html output)
interfere with diff's output on a deeper level, hiding it from all presenters.
This was surprising with --text output since it does not honour max-report-size
and is meant to be a "pure-by-default" output that shows everything.

So now we move this behaviour to a separate flag, max-diff-block-lines-saved,
and document it more clearly. Most users should not need this, unless they keep
running out of memory.