Skip the java tests if javap is older than 1.8
This fixes test_java and test_dex failing in jessie-backports.
These tests were failing because of slight behavior change between "javap"
tools versions.  The change was purely cosmetic, so no changes are made to
comparator itself, just the relevant tests are now skipped if "javap" is not
new enough.

See #838984 for more background.

Signed-off-by: Mattia Rizzolo <mattia@debian.org>