CONTRIBUTING: Refresh instructions for contributing to diffoscope.
The previous instructions suggested multiple ways to contribute.  Personally
I didn't feel sure which one was really right to use.  These encourage the
reader to send everything to the BTS, based on the conversation at
<http://meetbot.debian.net/reproducible-builds/2017/reproducible-builds.2017-01-03-18.00.log.html#l-38>.

I also tried to make the instructions a little less Debian-specific by
pointing people to the instructions for sending reports by e-mail, and then
mentioning reportbug as a helper for people who have it.

Signed-off-by: Mattia Rizzolo <mattia@debian.org>