Fix bug introduced in commit 36d1c964 that only worked "accidentally"
The previous behaviour (before the commit) returns False if neither regexes are
defined. It is not possible to reduce this to a simple boolean expression; one
needs to explicitly check for the corner-case "neither are defined". The
behaviour in the other cases was correct (both before and after the commit),
however the use of the "match-all" regexp makes it hard to express the corner
case. So, rewrite the whole thing.