Add missing Recommends for comparators. (Closes: #854655)
We ensure all comparators are imported so tool_required.all is populated;
this was a regression introduced in 4cdfa577f090ca942db42127811937357ae7135
as we no now lazily import our comparators.

This is difficult to test due to 5362315ec73eb5f6183bad20afe50f698cfa91f4
as we reload Comparators after every test so we are always in a consistent
state.

Signed-off-by: Chris Lamb <lamby@debian.org>